REF: Add tests.groupby.methods
