+++ pandas/tests/groupby/methods/test_any_all.py
