+++ pandas/tests/groupby/methods/test_groupby_shift_diff.py
