+++ pandas/tests/groupby/methods/test_sample.py
