+++ pandas/tests/groupby/methods/test_size.py
